clean up project and post detail pages
