--- app/project/page.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Head } from '../../components';
-
-const Page: React.FC = () => (
-    <main className="content">
-        <Head
-            title="Resume | Brian Behrens | Los Angeles Software Engineer"
-            description="The personal resume of Brian Behrens, Staff Software Engineer at Surfline."
-        />
-        testing
-    </main>
-);
-
-export default Page;
