+++ app/resume/page.tsx
@@ -5,7 +5,7 @@ import { type NextPage } from 'next';
 import { Head, ProfileHeader } from '../../components';
 
 const Page: NextPage = () => (
-    <main>
+    <main className="content">
         <Head
             title="Resume | Brian Behrens | Los Angeles Software Engineer"
             description="The personal resume of Brian Behrens, Staff Software Engineer at Surfline."
